refactor: manage directories, manage some code
